Check whether query exists before deleting it
The query delete view function now properly checks whether the query
that should be deleted exists and returns an error message if it
doesn't. The next step is to process the error message in the interface.

+++ query/views.py
@@ -88,7 +88,10 @@ def create_query(request):
 def delete(request, query_id):
     # TODO: check whether query belongs to the user
 
-    query = get_object_or_404(Query, pk=query_id)
+    query = Query.objects.get(pk=query_id)
+    if not query:
+        return json_response_message('ERROR', 'Query not found.')
+
     query.delete()
 
     return json_response_message('SUCCESS', '')
